+++ lib/screen/profile/alt/alt_profile_controller.dart
@@ -100,9 +100,13 @@ class AltProfileController extends BaseController {
   }
 
   void doLogout() async {
-    storage.deleteToken();
-    storage.deleteString(StorageCore.favoriteMenu);
-    Get.offAll(const LoginScreen());
+    await auth.logout().then((value){
+      if(value.code == 200){
+        storage.deleteToken();
+        storage.deleteString(StorageCore.favoriteMenu);
+        Get.offAll(const LoginScreen());
+      }
+    });
   }
 
   void isCcrfAction(dynamic screen, BuildContext context) {
